Add files via upload
The prediction of share value is one of the trickiest machine learning issues. In order to make wise judgments, investors must be able to quickly and accurately gather information about the stock market. Additionally, stock price behaviour is unpredictable and challenging to forecast. These factors make stock price forecasting both a crucial and difficult procedure. Finding the prediction model with the lowest error rate and highest forecast accuracy becomes a focus of research as a result. Time series data are used to represent stock prices, and neural networks are trained to identify patterns from trends in the historical data. To increase the accuracy of stock price prediction, this system used an LSTM algorithm.